+++ front-end/src/main.jsx
@@ -1,7 +1,7 @@
 
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Signup from "./pages/signup/signup"; // Signup 컴포넌트 불러오기
+import Signup from "./pages/auth/signup"; // Signup 컴포넌트 불러오기
 import "./style/scss/style.scss"; // 글로벌 스타일 적용
 import Favicon from "./component/logo/favicon"; 
 ReactDOM.createRoot(document.getElementById("root")).render(
